[IMP] mail: add default values coming from activity types when scheduling activities
Purpose of the commit is that in some cases people would like to add a
default user and a default description on their activity types. Those
default values will be propagated to created activities.

This commit adds fields to set the default description and default user on
activity type and propagate them through the various scheduling methods as
default values when the user did not give specific one.

Other usability improvements
* rename "Category" into "Action to Perform";
* change 'Other' value to 'None';
That way it is clearer that standard activities perform no action.

Related to task 1929742

closes odoo/odoo#30511

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>