[FIX] web: fix image placeholder size
Following commit 3ee3673af1013eb54f4de7be58cfecb94bb2b256

Before the current commit the placeholder was used at the end of the logic, so
it ignored the parameters such as width, height, crop, ...

Now the image is correctly replaced by the placeholder before we try to apply
the rest of the logic.

One example of such problem was in website_sale_comparison, on the comparison
page, if we tried to compare a product which had no image, the placeholder image
was bigger than expected.

task-1929646

closes odoo/odoo#30525